+++ tests/test_client.py
@@ -534,3 +534,23 @@ class ApiClientTest(unittest.TestCase):
         result = client.calculate_seconds_until_midnight(dt)
         expected_result = 3600 # 1 hour in seconds
         self.assertEqual(result, expected_result)
+
+    def test_user_has_privilege(self):
+        client = MarklogicApiClient("", "", "", False)
+
+        with patch.object(client, 'eval'):
+            user = "laura"
+            privilege_uri = "https://caselaw.nationalarchives.gov.uk/custom/uri"
+            privilege_action = "execute"
+            expected_vars = {
+                "user": "laura",
+                "privilege_uri": "https://caselaw.nationalarchives.gov.uk/custom/uri",
+                "privilege_action": "execute"
+            }
+            client.user_has_privilege(user, privilege_uri, privilege_action)
+
+            client.eval.assert_called_with(
+                os.path.join(ROOT_DIR, "xquery", "user_has_privilege.xqy"),
+                vars=json.dumps(expected_vars),
+                accept_header="application/xml"
+            )